requirements.txt
amqp==5.1.1
argcomplete==1.12.3
asgiref==3.5.2
astroid==2.11.5
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.23.9
botocore==1.26.10
celery==5.2.6
celery-progress==0.1.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commitizen==2.25.0
decli==0.5.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.4
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-import-export==2.8.0
django-storages==1.12.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
excel2json-3==0.1.6
exceptiongroup==1.0.0rc7
flake8==4.0.1
flower==1.0.0
future==0.18.2
generic==1.1.0
humanize==4.1.0
idna==3.3
importlib-resources==5.7.1
isort==5.10.1
Jinja2==3.1.2
jmespath==1.0.0
jsonschema==4.5.1
kombu==5.2.4
kwargs==1.0.1
lazy-object-proxy==1.7.1
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.6.1
Naked==0.1.31
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
platformdirs==2.5.2
postgres==4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.9
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
questionary==1.10.0
redis==4.3.1
requests==2.27.1
s3transfer==0.5.2
self==2020.12.3
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
status==0.2.5
tablib==3.2.1
termcolor==1.1.0
tomli==2.0.1
tomlkit==0.10.2
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.8.0
django-dotenv==1.4.2
